feat: Add source & source version attribution
This adds the source & source version attribution handling.

This makes it so that when configuring the SDK, the `source` and
`sourceVersion` can be configured and passed through as correlation
context for attribution on the backend.

This will make it so we can easily attribute AgentKit requests to
the specific version.